fix: Relocate API_VERSION global variable
Put 'API_VERSION' and 'set_api_version' static method in api.py.
This not only makes logical sense, but will prevent the circular
import preventing the SDK from being compatible with Python 3.6.
