+++ smartcar/__init__.py
@@ -1,9 +1,11 @@
 __version__ = "semantic-release"
 
+from smartcar.api import set_api_version
+
 from smartcar.auth_client import AuthClient
 
-from smartcar.static import set_api_version, get_user, get_vehicles, get_compatibility
+from smartcar.exception import SmartcarException
+
+from smartcar.static import get_user, get_vehicles, get_compatibility
 
 from smartcar.vehicle import Vehicle
-
-from smartcar.exception import SmartcarException
